files: rename and simplify components
